requirements.txt
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
cloudinary==1.17.0
dj-database-url==0.5.0
Django==2.2.16
django-cloudinary-storage==0.2.3
idna==2.10
mock==4.0.2
Pillow==7.2.0
pycodestyle==2.6.0
python-decouple==3.3
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
whitenoise==5.2.0
gunicorn
django-ckeditor
djangorestframework
